"Complete" MultiMC support
The -mmc option will now attempt to locate MultiMC on the system path
and use its directory as the MC home directory, or it can be specified
from the command line. It will also read the instance directory from
the MultiMC config and use that for relevant commands, such as
pack installation/creation. Also added support for reading the pack name
from the downloaded manifest during pack install